Split the transformation spec into helpers and regressions
The file had grown to six independent regression specs sharing a handful of
readers, and each new defect added a fixture path, a doc comment and a spec to
the same place. Spec.hs keeps the fixture-driven half and main, Spec.Regression
holds one spec per reproduced defect, Spec.Helpers the readers they share.
